refactor(plan-lifecycle): inline implementations, delete old modules
Consolidate the 5 underlying modules (plan-detection, frontmatter,
receipt, global-state, pipeline-state) into plan-lifecycle.ts,
replacing the re-export facade with real code.

Delete the old source modules and their test files. All behavioral
tests are preserved in the consolidated plan-lifecycle.test.ts.
Update ARCHITECTURE.md to reflect the new module structure.
